requirements.txt
click~=8.1.3
PyYAML~=6.0
Jinja2~=3.1.2
setuptools~=65.5.1
six~=1.16.0
rich~=13.6.0
Pygments~=2.16.1
pip~=22.3.1
attrs~=22.1.0
colorlog~=6.7.0
wheel~=0.38.4
Pillow~=10.0.1
pytest~=7.2.0
pluggy~=1.3.0
exceptiongroup~=1.0.4
iniconfig~=1.1.1
argcomplete~=2.0.0
packaging~=23.2
future~=0.18.3
mdurl~=0.1.2
MarkupSafe~=2.1.1
platformdirs~=3.11.0